requirements.txt
asgiref==3.5.2
Django==3.2
sqlparse==0.4.2
psycopg2==2.9.3
pandas==1.4.3
djangorestframework==3.13.1
django-filter==22.1
